+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and fill in the contents as you go. This simplifies later release management. --
 ## 2.3.0
 
 - Allow a plugin's entry-point file -- a single-file mu-plugin, or the main file in a plugin folder under `plugins/`, `mu-plugins/`, or `client-mu-plugins/` named `<dir>/<dir>.php` or `<dir>/plugin.php` -- to declare a namespace without being named `namespace.php` #343
+- Do not lint PHP within `build/` and `dist/` build directories #342
 
 ## 2.2.1